cast to float

+++ 02_MavenPlayground/src/main/java/at/htlle/pos/pocketCalculator.java
@@ -5,8 +5,8 @@ public class pocketCalculator {
     //basic claculator
     //+, -, *, /
 
-    public static int  addition(int numb1, int numb2){
-        return numb1 + numb2;
+    public static long addition(int numb1, int numb2){
+        return (long)numb1 + numb2;
     }
 
     public static int  subtraction(int numb1, int numb2){
